requirements.txt
asgiref==3.4.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.16
django-ckeditor==6.2.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
idna==3.3
mssql-django==1.1.3
openpyxl==3.0.9
Pillow
PyJWT==2.3.0
pyodbc
pytz==2021.1
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.9
